requirements.txt
requests
beautifulsoup4
pymupdf
google-search-python
tqdm 